Extract SignInForm into a separate component from SignIn screen
